renderer implemented

+++ LWDviz/src/LWDviz/ImGui/ImGuiLayer.cpp
@@ -70,7 +70,7 @@ namespace lw {
 	void ImGuiLayer::end() {
 		ImGuiIO& io = ImGui::GetIO();
 		Application& app = Application::get();
-		io.DisplaySize = ImVec2(app.getWindow().getWidth(), app.getWindow().getHeight());
+		io.DisplaySize = ImVec2((float)app.getWindow().getWidth(), (float)app.getWindow().getHeight());
 
 		// rendering
 		ImGui::Render();
